Address review: shared selector home, tz-aware zip name
- Move selection_ids (+ the MAX_SELECTION bound) out of the document
  adapter into adapters/_common — sibling adapters no longer import
  from one another.
- The bundle zip's date now uses the caller's timezone like the entry
  names inside it, so a near-midnight export can't produce
  document-2026-07-14.zip full of *-2026-07-15 entries. The worker
  passes the job selector's tz through render_to_storage.
